+++ packages/core/src/pie-chart.ts
@@ -108,7 +108,7 @@ export class PieChart extends BaseChart {
 			.attr("height", `${diameter}px`);
 
 		this.innerWrap
-			.style("transform", `translate(${radius}px,${radius}px)`)
+			.attr("transform", `translate(${radius},${radius})`)
 			.attr("width", `${diameter}px`)
 			.attr("height", `${diameter}px`)
 			.attr("preserveAspectRatio", "xMinYMin");